🎨 POST, PUT, DELETE用のメソッドを1つのhooksにしてsnackbarを設定
